Suppress MSVC "warning C4267: 'argument' : conversion from 'size_t' to 'int', possible loss of data" (introduced in commit:1ec131d3)
(cherry picked from commit ab36b3a225ae22409d4b7bba121f8a0fbae95311)

+++ Development/mdns/bonjour_dns_impl.cpp
@@ -232,7 +232,7 @@ PRAGMA_WARNING_POP
     inline std::string ierase_tail_copy(const std::string& input, const std::string& tail)
     {
         return boost::algorithm::iends_with(input, tail)
-            ? boost::algorithm::erase_tail_copy(input, tail.size())
+            ? boost::algorithm::erase_tail_copy(input, (int)tail.size())
             : input;
     }
